Moved and renamed auth helper
